Initial upload of test test code for MS5803-14BA
This code displays pressure and temperature values from the MS5803-14BA
detected through its I2C bus on the "serial monitor" screen in the
Arduino IDE.  Because the MS5803-14BA uses a 3v3 logic, the user should
take care to only power the devise with 3v3 (not 5v).  CLK and SDA seem
to work when attached directly to the Arduino, however.